Partition correctness tests
Clean up log
