+++ test/Operators/finitedifference/linsolve.jl
@@ -1,9 +1,11 @@
+#=
+julia --project=.buildkite
+using Revise; include(joinpath("test", "Operators", "finitedifference", "linsolve.jl"))
+=#
 using Test
 using ClimaComms
-
+ClimaComms.@import_required_backends
 import ClimaCore
-# To avoid JET failures in the error message
-ClimaCore.Operators.allow_mismatched_fd_spaces() = true
 
 using ClimaCore:
     Geometry, Domains, Meshes, Topologies, Spaces, Fields, Quadratures
@@ -89,11 +91,9 @@ W = SchurComplementW(Y, use_transform, jacobi_flags)
 
 using JET
 using Test
-@time test_linsolve!(Y, W, b)
-@time test_linsolve!(Y, W, b)
 
 @testset "JET test for `apply` in linsolve! kernel" begin
+    test_linsolve!(Y, W, b) # compile first
+    @test 0 == @allocated test_linsolve!(Y, W, b)
     @test_opt test_linsolve!(Y, W, b)
 end
-
-ClimaCore.Operators.allow_mismatched_fd_spaces() = false
